LOADER: move the logic from `LibrarySccTycker` to `LibraryModuleLoader`
